+++ battlecode22-scaffold/CHANGELOG.md
@@ -85,3 +85,4 @@ than max rubble.
 * Sage will not quickly run to archon now when action on cooldown, will wait for enemies in range.
 * Created Anomaly.java
 * Modified BotMiner: increased findOptimalMiningLocation(lead) lower bound to 8; Decreased ratio (dist to allied archon/ enemy archon) of mine depletion check to 1.2
+* Laboratory now moves to adjacent lesser rubble area if present.